refactor: 같은 패키지 내에 할인 도메인 관리
