+++ src/main/java/christmas/domain/discount/WeekendDiscountPolicy.java
@@ -1,4 +1,9 @@
-package christmas.domain;
+package christmas.domain.discount;
+
+import christmas.domain.DayChecker;
+import christmas.domain.FoodCategory;
+import christmas.domain.Order;
+import christmas.domain.Orders;
 
 import java.time.LocalDate;
